requirements.txt
streamlit==1.29.0
google-genai==0.3.0
PyPDF2==3.0.1
python-docx==1.1.0
plotly==5.17.0
pandas==2.1.4
numpy>=1.21.0